Improve graph generation: create separate GraphGenerator, implement non-tree generation. Merge ModelGenerator inside DungeonGenerator module. Collect all generation settings inside separate Settings.h.
