--- src/model-generator/CMakeLists.txt
@@ -1,20 +1,0 @@
-cmake_minimum_required(VERSION 3.23)
-
-project(model_generator)
-add_library(${PROJECT_NAME} STATIC
-    ModelGenerator.cpp
-)
-
-target_sources(${PROJECT_NAME} PUBLIC
-    FILE_SET "${PROJECT_NAME}_HEADERS"
-    TYPE HEADERS
-    BASE_DIRS
-        "."
-    FILES
-        "ModelGenerator.h"
-)
-
-target_link_libraries(${PROJECT_NAME}
-    PUBLIC
-        model
-)
